requirements.txt
Django==5.0.1
python-dotenv==1.0.0
google-cloud-aiplatform==1.38.1
Pillow==10.2.0
google-generativeai==0.3.2
python-magic==0.4.27